+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org).
 
 ## [Unreleased]
 
+## [0.5.0] - 2026-08-21
+
 ### Fixed
 
 - Prevented interactive progress spinners from printing every animation frame when